Pre-apply counter effect images as masks of opaque white.
Mix blend modes only blend colors with element the webpage. In
Streamlabs OBS, the webpage background is essentially transparent,
so other widgets underneath the webpage are visible but not blended
with the color of the masks. The result is that black is rendered
on the transparent background on the webpage, which is not intended.