Added tetris, putet nuisance styles.
Fixed a bug where images are actually bigger than they should be.
This is due to specifying width to be 95%. The aspect ratio technique
uses percentage values for padding-bottom, and this percentage is
with respect to the element's container. Having width: 95% on the
element itself will mess up the aspect ratio.