Fixed possibility of comet image overflowing nuisance queue.
We are leaving only very little space on both sides of the screen,
but the comet image can well overflow width / 6 space so it can
be off screen. The result is that offscreen portion of the image
won't be rendered in StreamLabs, which is undesirable.

Adding extra margin on the right side of the screen to avoid queue
being cut off.